evaluation/requirements.txt
# Ragas 评测所需依赖（按 ragas 官方文档 + 自身项目集成需要）
# 说明：
#   * ragas 与 datasets / langchain 深度耦合，按官方要求安装
#   * 用 langchain-openai 去对接 SiliconFlow / DashScope / OpenAI 等兼容 OpenAI 的接口
#   * python-dotenv 用来加载 .env 里的密钥与后端地址
#   * httpx 用来调用自己的 Spring Boot 评测端点（登录 + 评测）
#   * pandas 用来落地结果 CSV 方便查看
ragas>=0.3.6
datasets>=3.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.1
httpx>=0.27.0
pandas>=2.2.0
tqdm>=4.66.0
